requirements.txt
apache-airflow==2.4.1
google-cloud-bigquery==3.3.3
pandas==1.5.0
dbt-bigquery==1.2.0
